WiP: Major changes
Added a system for future use in which the main programme and renderer logic are put into a tree-based structure. This allows for a system where most of both initialisation/deinitialisation and update logic can be called from the same tree structure. More information on the details will be written later.

As of current, it's still very much a work in progress and does not run properly. It does compile.