Texture utils: Improved bmp-dds conversions
Simplified algorithms for loading and converting bmp/dib and dds formats. Rewrote a large part of bmp->dds conversion to produce a better result. There are still a lot of improvements to be done in that respect (as there often are when it comes to lossy algorithms), but it produces images that look rather decent and works also with images that are more artificial and less organic.

In addition to changes in functionality, added more detailed comments about the algorithms.